Stop at 3 hours
        - My main goal with the three hours alloted was to understand
the question being asked and the nuances it presented. Thus, this work is
exploritory.
        - Next steps would likely be based on the intended use of this
code. If the code was intended to be used by another engineer or to be
reused with other lists of zip codes I would likely delete this work and
build something with small, easily testable classes. However, if this
was a simple request for data I would likely leave it as is.